+++ packages/react-devtools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @assistant-ui/react-devtools
 
+## 1.2.0
+
+### Patch Changes
+
+- Updated dependencies [[`040d469`](https://github.com/assistant-ui/assistant-ui/commit/040d469acfcf782de6fc188c646dfd8732d27088)]:
+  - @assistant-ui/react@0.14.0
+
 ## 1.1.0
 
 ### Patch Changes