+++ packages/react-ink/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @assistant-ui/react-ink
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [[`040d469`](https://github.com/assistant-ui/assistant-ui/commit/040d469acfcf782de6fc188c646dfd8732d27088)]:
+  - @assistant-ui/core@0.2.0
+
 ## 0.0.14
 
 ### Patch Changes