+++ packages/react-lexical/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @assistant-ui/react-lexical
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies [[`040d469`](https://github.com/assistant-ui/assistant-ui/commit/040d469acfcf782de6fc188c646dfd8732d27088)]:
+  - @assistant-ui/react@0.14.0
+  - @assistant-ui/core@0.2.0
+
 ## 0.1.0
 
 ### Patch Changes